MxOnline/requirements.txt
certifi==2019.9.11
chardet==3.0.4
defusedxml==0.6.0
diff-match-patch==20181111
Django==2.2.7
django-crispy-forms==1.8.0
django-formtools==2.1
django-import-export==1.2.0
django-pure-pagination==0.3.0
django-ranged-response==0.2.0
django-reversion==3.0.4
django-simple-captcha==0.5.12
djangorestframework==3.10.3
docopt==0.6.2
et-xmlfile==1.0.1
future==0.18.2
httplib2==0.14.0
idna==2.8
jdcal==1.4.1
MarkupPy==1.14
odfpy==1.4.0
openpyxl==3.0.0
Pillow==6.2.1
pipreqs==0.4.9
progressbar2==3.47.0
PyMySQL==0.9.3
python-utils==2.3.0
pytz==2019.3
PyYAML==5.1.2
requests==2.22.0
six==1.13.0
sqlparse==0.3.0
tablib==0.14.0
urllib3==1.25.6
xlrd==1.2.0
XlsxWriter==1.2.5
xlwt==1.3.0
yarg==0.1.9
